# [C++] Drastically improve multi-threaded performance