--- a/runtime/Cpp/runtime/src/atn/ATN.cpp
+++ b/runtime/Cpp/runtime/src/atn/ATN.cpp
@@ -20,65 +20,14 @@ using namespace antlr4;
 using namespace antlr4::atn;
 using namespace antlrcpp;
 
-ATN::ATN() : ATN(ATNType::LEXER, 0) {
-}
+ATN::ATN() : ATN(ATNType::LEXER, 0) {}
 
-ATN::ATN(ATN &&other) {
-  // All source vectors are implicitly cleared by the moves.
-  states = std::move(other.states);
-  decisionToState = std::move(other.decisionToState);
-  ruleToStartState = std::move(other.ruleToStartState);
-  ruleToStopState = std::move(other.ruleToStopState);
-  grammarType = std::move(other.grammarType);
-  maxTokenType = std::move(other.maxTokenType);
-  ruleToTokenType = std::move(other.ruleToTokenType);
-  lexerActions = std::move(other.lexerActions);
-  modeToStartState = std::move(other.modeToStartState);
-}
-
-ATN::ATN(ATNType grammarType_, size_t maxTokenType_) : grammarType(grammarType_), maxTokenType(maxTokenType_) {
-}
+ATN::ATN(ATNType grammarType_, size_t maxTokenType_) : grammarType(grammarType_), maxTokenType(maxTokenType_) {}
 
 ATN::~ATN() {
   for (ATNState *state : states) {
     delete state;
   }
-}
-
-/**
- * Required to be defined (even though not used) as we have an explicit move assignment operator.
- */
-ATN& ATN::operator = (ATN &other) noexcept {
-  states = other.states;
-  decisionToState = other.decisionToState;
-  ruleToStartState = other.ruleToStartState;
-  ruleToStopState = other.ruleToStopState;
-  grammarType = other.grammarType;
-  maxTokenType = other.maxTokenType;
-  ruleToTokenType = other.ruleToTokenType;
-  lexerActions = other.lexerActions;
-  modeToStartState = other.modeToStartState;
-
-  return *this;
-}
-
-/**
- * Explicit move assignment operator to make this the preferred assignment. With implicit copy/move assignment
- * operators it seems the copy operator is preferred causing trouble when releasing the allocated ATNState instances.
- */
-ATN& ATN::operator = (ATN &&other) noexcept {
-  // All source vectors are implicitly cleared by the moves.
-  states = std::move(other.states);
-  decisionToState = std::move(other.decisionToState);
-  ruleToStartState = std::move(other.ruleToStartState);
-  ruleToStopState = std::move(other.ruleToStopState);
-  grammarType = std::move(other.grammarType);
-  maxTokenType = std::move(other.maxTokenType);
-  ruleToTokenType = std::move(other.ruleToTokenType);
-  lexerActions = std::move(other.lexerActions);
-  modeToStartState = std::move(other.modeToStartState);
-
-  return *this;
 }
 
 misc::IntervalSet ATN::nextTokens(ATNState *s, RuleContext *ctx) const {
@@ -89,7 +38,7 @@ misc::IntervalSet ATN::nextTokens(ATNState *s, RuleContext *ctx) const {
 
 misc::IntervalSet const& ATN::nextTokens(ATNState *s) const {
   if (!s->_nextTokenUpdated) {
-    std::unique_lock<std::mutex> lock { _mutex };
+    std::unique_lock<std::mutex> lock(_mutex);
     if (!s->_nextTokenUpdated) {
       s->_nextTokenWithinRule = nextTokens(s, nullptr);
       s->_nextTokenUpdated = true;