--- a/runtime/Cpp/runtime/src/atn/ATNSimulator.h
+++ b/runtime/Cpp/runtime/src/atn/ATNSimulator.h
@@ -41,9 +41,6 @@ namespace atn {
     virtual Ref<const PredictionContext> getCachedContext(Ref<const PredictionContext> const& context);
 
   protected:
-    static std::shared_mutex _stateLock; // Lock for DFA states.
-    static std::shared_mutex _edgeLock; // Lock for the sparse edge map in DFA states.
-
     /// <summary>
     /// The context cache maps all PredictionContext objects that are equals()
     ///  to a single cached copy. This cache is shared across all contexts