--- a/runtime/Cpp/runtime/src/misc/InterpreterDataReader.h
+++ b/runtime/Cpp/runtime/src/misc/InterpreterDataReader.h
@@ -13,7 +13,7 @@ namespace antlr4 {
 namespace misc {
 
   struct InterpreterData {
-    atn::ATN atn;
+    std::unique_ptr<atn::ATN> atn;
     dfa::Vocabulary vocabulary;
     std::vector<std::string> ruleNames;
     std::vector<std::string> channels; // Only valid for lexer grammars.