… runtime/Cpp/runtime/src/atn/ParserATNSimulator.cpp
… runtime/Cpp/runtime/src/tree/xpath/XPathLexer.cpp
… runtime/Cpp/runtime/src/tree/xpath/XPathLexer.h
… tool/resources/org/antlr/v4/tool/templates/codegen/Cpp/Cpp.stg