The original implementation uses two static `std::shared_mutex` that are shared among all `ParserATNSimulator` and `LexerATNSimulator`. This causes all generated lexers and parsers to share the same global locks. This is unnecessary as all data is rooted in `antlr4::atn::ATN`, meaning all data that needs synchronized access belongs to exactly one `antlr4::atn::ATN`. With this in mind, we can shift the locks from static to per `antlr4::atn::ATN`. This drastically improves performance as the number of threads go up, in some cases as much as 40%.