@parrt 2022-02-24
If this PR proves/looks useful, @mike-lischke , I'd like to wait to incorporate this into 4.10.